+++ src/Stormancer.Plugins/GameVersion/Stormancer.Server.Plugins.GameVersion/Changelog.rst
@@ -14,7 +14,9 @@ Unreleased
 Fixed
 *****
 - Fixed possible null reference exception if the configuration was not set.
-
+Added
+*****
+- Added TryGetVersion extension method to get the version field from a session without knowing the field name.
 2.3.0
 ----------
 Changed